+++ src/core/meta_commands.rb
@@ -28,6 +28,30 @@ module Fizzy::MetaCommands
             git_clone(spec["repo"], spec["dst"])
           end
         }
+      },
+      "download" => {
+        "validator" => lambda { |spec|
+          # 1. Validate.
+          status   = spec.has_key?("url")
+          status &&= spec.has_key?("dst")
+          # 2. Normalize.
+          if status
+            spec["url"] = URI(spec["url"])
+            spec["dst"] = Pathname.new(spec["dst"]).expand_path
+          end
+          status
+        },
+        "executor" => lambda { |spec|
+          res = Net::HTTP.get_response(spec["url"])
+          if res.is_a?(Net::HTTPSuccess)
+            # TODO atm it requires the current user has write access.
+            #      refactor when a more robust permission mgmt is implemented.
+            FileUtils.mkdir_p(spec["dst"].dirname)
+            spec["dst"].write(res.body)
+          else
+            error("Network error: cannot retrieve `#{spec["url"]}`.")
+          end
+        }
       }
     }
   end